Remove conversion

+++ packages/orchestrator/internal/sandbox/uffd/userfaultfd/serve.go
@@ -36,7 +36,7 @@ func (u *userfaultfd) Serve(
 ) error {
 	pollFds := []unix.PollFd{
 		{Fd: int32(u.fd), Events: unix.POLLIN},
-		{Fd: int32(fdExit.Reader()), Events: unix.POLLIN},
+		{Fd: fdExit.Reader(), Events: unix.POLLIN},
 	}
 
 	var eg errgroup.Group
